requirements.txt
gradio
easyocr
opencv-python
numpy
python-dateutil


